requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.1
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Mako==1.3.9
MarkupSafe==3.0.2
orjson==3.10.15
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pytest==8.3.4
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
slugify==0.0.1
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
text-unidecode==1.3
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
